changed location of command and fixed syntax
