+++ mcodingbot/plugins/info.py
@@ -5,7 +5,6 @@ import hikari
 
 from mcodingbot.config import CONFIG
 from mcodingbot.utils import Context, Plugin
-from mcodingbot.plugins import stats
 
 plugin = Plugin()
 
@@ -29,19 +28,3 @@ async def links(ctx: Context) -> None:
         ),
     )
     await ctx.respond(embed=embed)
-
-@plugin.include
-@crescent.command(name="stats", description="Exact values for mCoding statistics")
-async def stats(ctx: Context) -> None:
-    stats = await stats.get_stats(ctx.session.bot)
-    embed = hikari.Embed(
-        title="mCoding stats",
-        color=CONFIG.theme,
-        description=(
-            f"Server members: `{ctx.server.member_count:,}`\n"
-            f"Subscribers: `{stats.subs:,}`\n"
-            f"Views: {stats.views:,}"
-        ),
-    )
-    await ctx.respond(embed=embed)
-
